Move essential parts of the RPN execution to the RPN class.
